Added Jacoco and dependency check, adjust the tests
