+++ Jdempotent-core/src/test/java/com/trendyol/jdempotent/core/aspect/TestAopContext.java
@@ -1,4 +1,4 @@
-package aspect.withaspect;
+package com.trendyol.jdempotent.core.aspect;
 
 import com.trendyol.jdempotent.core.aspect.IdempotentAspect;
 import com.trendyol.jdempotent.core.datasource.InMemoryIdempotentRepository;
