+++ Jdempotent-core/src/test/java/com/trendyol/jdempotent/core/chain/JdempotentDefaultChainTest.java
@@ -1,8 +1,6 @@
-package aspect.chain;
+package com.trendyol.jdempotent.core.chain;
 
-import aspect.core.IdempotentTestPayload;
-import com.trendyol.jdempotent.core.chain.JdempotentDefaultChain;
-import com.trendyol.jdempotent.core.chain.JdempotentPropertyAnnotationChain;
+import com.trendyol.jdempotent.core.utils.IdempotentTestPayload;
 import com.trendyol.jdempotent.core.model.ChainData;
 import com.trendyol.jdempotent.core.model.KeyValuePair;
 import org.junit.Test;
@@ -13,26 +11,26 @@ import org.springframework.test.context.junit4.SpringRunner;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(SpringRunner.class)
-public class JdempotentPropertyAnnotationChainTest {
+public class JdempotentDefaultChainTest {
 
     @InjectMocks
-    private JdempotentPropertyAnnotationChain jdempotentPropertyAnnotationChain;
+    private JdempotentDefaultChain jdempotentDefaultChain;
 
     @Test
     public void should_process_with_no_annotation() throws IllegalAccessException, NoSuchFieldException {
         //Given
         IdempotentTestPayload idempotentTestPayload = new IdempotentTestPayload();
-        idempotentTestPayload.setEventId(1l);
+        idempotentTestPayload.setName("value");
         ChainData chainData = new ChainData();
         chainData.setArgs(idempotentTestPayload);
-        chainData.setDeclaredField(idempotentTestPayload.getClass().getDeclaredField("eventId"));
+        chainData.setDeclaredField(idempotentTestPayload.getClass().getDeclaredField("name"));
 
         //When
-        KeyValuePair process = jdempotentPropertyAnnotationChain.process(chainData);
+        KeyValuePair process = jdempotentDefaultChain.process(chainData);
 
         //Then
-        assertEquals("transactionId", process.getKey());
-        assertEquals(1l, process.getValue());
+        assertEquals("name", process.getKey());
+        assertEquals("value", process.getValue());
     }
 
     @Test
@@ -45,10 +43,10 @@ public class JdempotentPropertyAnnotationChainTest {
         chainData.setDeclaredField(idempotentTestPayload.getClass().getDeclaredField("eventId"));
 
         //When
-        KeyValuePair process = jdempotentPropertyAnnotationChain.process(chainData);
+        KeyValuePair process = jdempotentDefaultChain.process(chainData);
 
         //Then
-        assertEquals("transactionId", process.getKey());
+        assertEquals("eventId", process.getKey());
         assertEquals(1l, process.getValue());
     }
 }
