+++ Jdempotent-core/src/test/java/com/trendyol/jdempotent/core/chain/JdempotentPropertyAnnotationChainTest.java
@@ -1,7 +1,6 @@
-package aspect.chain;
+package com.trendyol.jdempotent.core.chain;
 
-import aspect.core.IdempotentTestPayload;
-import com.trendyol.jdempotent.core.chain.JdempotentDefaultChain;
+import com.trendyol.jdempotent.core.utils.IdempotentTestPayload;
 import com.trendyol.jdempotent.core.model.ChainData;
 import com.trendyol.jdempotent.core.model.KeyValuePair;
 import org.junit.Test;
@@ -12,26 +11,26 @@ import org.springframework.test.context.junit4.SpringRunner;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(SpringRunner.class)
-public class JdempotentDefaultChainTest {
+public class JdempotentPropertyAnnotationChainTest {
 
     @InjectMocks
-    private JdempotentDefaultChain jdempotentDefaultChain;
+    private JdempotentPropertyAnnotationChain jdempotentPropertyAnnotationChain;
 
     @Test
     public void should_process_with_no_annotation() throws IllegalAccessException, NoSuchFieldException {
         //Given
         IdempotentTestPayload idempotentTestPayload = new IdempotentTestPayload();
-        idempotentTestPayload.setName("value");
+        idempotentTestPayload.setEventId(1l);
         ChainData chainData = new ChainData();
         chainData.setArgs(idempotentTestPayload);
-        chainData.setDeclaredField(idempotentTestPayload.getClass().getDeclaredField("name"));
+        chainData.setDeclaredField(idempotentTestPayload.getClass().getDeclaredField("eventId"));
 
         //When
-        KeyValuePair process = jdempotentDefaultChain.process(chainData);
+        KeyValuePair process = jdempotentPropertyAnnotationChain.process(chainData);
 
         //Then
-        assertEquals("name", process.getKey());
-        assertEquals("value", process.getValue());
+        assertEquals("transactionId", process.getKey());
+        assertEquals(1l, process.getValue());
     }
 
     @Test
@@ -44,10 +43,10 @@ public class JdempotentDefaultChainTest {
         chainData.setDeclaredField(idempotentTestPayload.getClass().getDeclaredField("eventId"));
 
         //When
-        KeyValuePair process = jdempotentDefaultChain.process(chainData);
+        KeyValuePair process = jdempotentPropertyAnnotationChain.process(chainData);
 
         //Then
-        assertEquals("eventId", process.getKey());
+        assertEquals("transactionId", process.getKey());
         assertEquals(1l, process.getValue());
     }
 }
