+++ Jdempotent-spring-boot-redis-starter/pom.xml
@@ -44,8 +44,8 @@
 
     <properties>
         <java.version>11</java.version>
-        <spring.boot.version>2.3.4.RELEASE</spring.boot.version>
-        <spring.framework.version>5.2.9.RELEASE</spring.framework.version>
+        <spring.boot.version>2.7.3</spring.boot.version>
+        <spring.framework.version>5.3.22</spring.framework.version>
         <version.mockito>1.10.19</version.mockito>
         <version.junit>4.13.1</version.junit>
     </properties>
@@ -64,7 +64,7 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-data-redis</artifactId>
-            <version>2.2.5.RELEASE</version>
+            <version>${spring.boot.version}</version>
             <exclusions>
                 <exclusion>
                     <groupId>org.springframework.boot</groupId>
@@ -83,13 +83,13 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.13.1</version>
+            <version>4.13.2</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.junit.vintage</groupId>
             <artifactId>junit-vintage-engine</artifactId>
-            <version>5.7.0</version>
+            <version>5.9.0</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -271,7 +271,7 @@
                 </plugin>
                 <plugin>
                     <artifactId>maven-surefire-plugin</artifactId>
-                    <version>2.22.1</version>
+                    <version>3.0.0-M7</version>
                 </plugin>
                 <plugin>
                     <artifactId>maven-jar-plugin</artifactId>
